Task 2 completion
This is the actual final branch commit before merging back to main. As noted in previous branch commit, the ideal  launch date and time have been determined for the transfer orbit. Task 2 is completed, with the appropriate vectors and anomaly angles displayed. Some code scraps were put under Task 3 subsection as they may be useful for it. This commit also implements some display changes for ease of reading outputs.