Mars Elliptical Orbit Update
Accounts for perihelion and aphelion of Mars orbit for worst-case scenarios rather than assuming circular orbit. Adjustment to paint fraction and ADCS were made. Overall mass is above the stated deliverable payload of a Falcon Heavy, so we will either need to figure out the C3 of a Saturn V (if it would possibly be enough), or just assume circular and ignore it.